Simplify Ops.
