+++ src/SpaceLiDAR.jl
@@ -71,7 +71,7 @@ export to_egm2008, to_egm2008!
 export topex_to_wgs84, topex_to_wgs84!
 export icesat_saturation_correct, icesat_saturation_correct!
 export icesat_quality
-export Operation, apply, apply!, inputs, outputs
+export Operation, Filter, Transform
 export ToEGM2008, TopexToWGS84, SaturationCorrect, ICESatQuality, InExtent
 export ICESat_Granule, ICESat2_Granule, GEDI_Granule, convert
 export granule, granules
