+++ src/operations.jl
@@ -1,20 +1,22 @@
 # ─── Operations: declarative filters and column transforms ───────────────────
 #
-# An `Operation` is a value that declares which columns it needs (`inputs`) and
-# which it produces (`outputs`), and knows how to either mask rows (a *filter*,
-# `outputs` empty) or overwrite columns (a *transform*, `outputs` non-empty).
+# An `Operation` declares which columns it needs (`_inputs`) and is either a
+# `Filter` (masks rows) or a `Transform` (overwrites columns). The kind is
+# encoded in the type hierarchy, not a runtime flag:
+#
+#   abstract type Filter    <: Operation end   # implements `_mask(op, cols)`
+#   abstract type Transform <: Operation end   # implements `_run!(op, cols)`
 #
 # Why this exists: the generic `H5Table` reader is decoupled from product
 # semantics, so post-processing used to probe columns at runtime
 # (`hasproperty`) and silently depend on the user having selected the right
 # variables. Operations turn that implicit contract explicit:
 #
-#   * `inputs(op, granule)` lists required columns as `Variable` specs — used to
+#   * `_inputs(op, granule)` lists required columns as `Variable` specs — used to
 #     *auto-pull* missing columns from a lazy `H5Table` (via `resolve_variable`
-#     on its source) so the user can't under-select. Auto-pulled columns that
-#     the op only needed transiently (e.g. lat/lon for a bbox filter) are
-#     dropped from the result.
-#   * `inputs` is *granule-dispatched*: generic operations specialise on the
+#     on its source) so the user can't under-select. Auto-pulled columns are
+#     kept in the result.
+#   * `_inputs` is *granule-dispatched*: generic operations specialise on the
 #     abstract `Granule`, product-bound ones on their concrete granule (e.g.
 #     `TopexToWGS84` only on `ICESat_Granule`). Inapplicable `(op, granule)`
 #     pairs hit the default method, which throws an applicability error.
@@ -25,9 +27,13 @@
 # the verbs and the auto-pull machinery; products are included last so this is
 # all defined before any product subtypes `Operation`.
 #
-# Verbs (mirroring the mutating/non-mutating split used elsewhere):
-#   filter(op, t) / filter!(op, t)   — row filters → subset rows
-#   apply(op, t)  / apply!(op, t)    — column transforms → overwrite columns
+# Only the operation *types* are public; the interface functions
+# (`_inputs`/`_mask`/`_run!`) are package-internal.
+#
+# Verbs (extending Base, mirroring the mutating/non-mutating split):
+#   filter(op, t) / filter!(op, t)   — `Filter` ops → subset rows
+#   map(op, t)    / map!(op, t)      — `Transform` ops → overwrite columns
+# Using the wrong verb for an op's kind is a `MethodError`.
 #
 # The non-mutating verbs accept lazy `H5Table`/`PartitionedH5Table` (auto-pull +
 # materialize) or any materialized table. The mutating verbs require an
@@ -36,7 +42,23 @@
 abstract type Operation end
 
 """
-    inputs(op::Operation, granule) -> Vector{Variable}
+    abstract type Filter <: Operation
+
+An [`Operation`](@ref) that masks rows. Used via [`filter`](@ref)/[`filter!`].
+Implements the internal `_mask(op, cols)::BitVector`.
+"""
+abstract type Filter <: Operation end
+
+"""
+    abstract type Transform <: Operation
+
+An [`Operation`](@ref) that overwrites columns. Used via [`map`](@ref)/[`map!`].
+Implements the internal `_run!(op, cols)`.
+"""
+abstract type Transform <: Operation end
+
+"""
+    _inputs(op::Operation, granule) -> Vector{Variable}
 
 Columns `op` reads, as full [`Variable`](@ref) specs. Granule-dispatched: the
 default method throws, so an operation only applies to granules it has a method
@@ -44,18 +66,10 @@ for. Generic operations specialise on `Granule`, product-bound ones on their
 concrete granule type. Pass `nothing` for a sourceless table (the op then
 returns name-only specs used purely for validation).
 """
-inputs(op::Operation, granule) = throw(ArgumentError(
+_inputs(op::Operation, granule) = throw(ArgumentError(
     "$(typeof(op)) is not applicable to " *
     (granule === nothing ? "a sourceless table" : string(typeof(granule))),
 ))
-
-"""
-    outputs(op::Operation) -> Vector{Symbol}
-
-Columns `op` writes. Empty marks `op` as a *filter* (row mask); non-empty marks
-it a *transform* (overwrites the listed columns).
-"""
-outputs(::Operation) = Symbol[]
 
 # ─── input-spec helpers (shared by generic operations) ────────────────────────
 
@@ -108,7 +122,7 @@ _containers(t::DataFrame) = (t,)
 _colnames(t) = Tables.columnnames(Tables.columns(t))
 
 # Names required by `op` for `granule`, used to validate materialized tables.
-_input_names(op::Operation, granule) = Symbol[v.name for v in inputs(op, granule)]
+_input_names(op::Operation, granule) = Symbol[v.name for v in _inputs(op, granule)]
 
 function _missing_col_msg(name::Symbol, granule)
     if granule === nothing
@@ -125,7 +139,7 @@ end
 function _validate(t, need)
     have = Set(_colnames(t))
     for name in need
-        name in have || error(_missing_col_msg(name, _opgranule(t)))
+        name in have || throw(ArgumentError(_missing_col_msg(name, _opgranule(t))))
     end
     return t
 end
@@ -136,120 +150,83 @@ function _augment(t::H5Tables.H5Table, need::Vector{Variable})
     have = Set(Tables.columnnames(t))
     src = getfield(t, :f)
     newvars = copy(t.vars)
-    transient = Symbol[]
     for v in need
         v.name in have && continue
         any(a -> a.name == v.name, t.attrs) && continue
         rv = H5Tables.resolve_variable(src, v)
-        rv === nothing && error(_missing_col_msg(v.name, _opgranule(t)))
+        rv === nothing && throw(ArgumentError(_missing_col_msg(v.name, _opgranule(t))))
         push!(newvars, rv)
-        push!(transient, v.name)
     end
-    H5Tables.H5Table(f = src, vars = newvars, attrs = t.attrs, nrow = t.nrow), transient
+    H5Tables.H5Table(f = src, vars = newvars, attrs = t.attrs, nrow = t.nrow)
 end
 
 function _augment(t::H5Tables.PartitionedH5Table, need::Vector{Variable})
-    augmented = H5Tables.H5Table[]
-    transient = Symbol[]
-    for p in t.tables
-        ap, transient = _augment(p, need)
-        push!(augmented, ap)
-    end
-    H5Tables.PartitionedH5Table(augmented), transient
+    H5Tables.PartitionedH5Table([_augment(p, need) for p in t.tables])
 end
 
-# Materialize `t` with the columns `need` available, returning the materialized
-# table and the names of columns that were pulled in only transiently.
-function _materialize_for(t::H5Tables.H5Table, need::Vector{Variable})
-    aug, tr = _augment(t, need)
-    return collect(aug), tr
-end
-function _materialize_for(t::H5Tables.PartitionedH5Table, need::Vector{Variable})
-    aug, tr = _augment(t, need)
-    return collect(aug), tr
-end
+# Materialize `t` with the columns `need` available.
+_materialize_for(t::H5Tables.H5Table, need::Vector{Variable}) = collect(_augment(t, need))
+_materialize_for(t::H5Tables.PartitionedH5Table, need::Vector{Variable}) =
+    collect(_augment(t, need))
 _materialize_for(t::AbstractTable, need::Vector{Variable}) =
-    (_validate(t, (v.name for v in need)); (_copytable(t), Symbol[]))
+    (_validate(t, (v.name for v in need)); _copytable(t))
 _materialize_for(t, need::Vector{Variable}) =
-    (_validate(t, (v.name for v in need)); (_copytable(t), Symbol[]))
+    (_validate(t, (v.name for v in need)); _copytable(t))
 
 _copytable(t::Table) = Table(map(copy, _table(t)), _granule(t))
 _copytable(t::PartitionedTable) = PartitionedTable(map(nt -> map(copy, nt), t.tables), _granule(t))
 _copytable(t::DataFrame) = copy(t)
 
-# ─── dropping transient columns ───────────────────────────────────────────────
-
-_drop_transient(t, transient) = isempty(transient) ? t : _dropcols(t, transient)
-function _dropcols(t::Table, names)
-    Table(Base.structdiff(_table(t), NamedTuple{Tuple(names)}), _granule(t))
-end
-function _dropcols(t::PartitionedTable, names)
-    PartitionedTable(map(nt -> Base.structdiff(nt, NamedTuple{Tuple(names)}), t.tables), _granule(t))
-end
-
 # ─── public verbs ─────────────────────────────────────────────────────────────
 
 """
-    apply(op::Operation, t)
+    map(op::Transform, t)
 
-Apply a transform `op` to table `t`, returning a new table with `outputs(op)`
-overwritten. For a lazy `H5Table`/`PartitionedH5Table`, missing input columns
-are auto-pulled from the granule source (and dropped again if only needed
-transiently). See [`apply!`](@ref) for the mutating version.
+Apply a transform `op` to table `t`, returning a new table with the transformed
+columns overwritten. For a lazy `H5Table`/`PartitionedH5Table`, missing input
+columns are auto-pulled from the granule source and kept in the result. See
+[`map!`](@ref) for the mutating version.
 """
-function apply(op::Operation, t::OpTable)
-    isempty(outputs(op)) &&
-        throw(ArgumentError("$(typeof(op)) is a filter; use `filter(op, t)`."))
-    g = _opgranule(t)
-    mt, transient = _materialize_for(t, inputs(op, g))
+function Base.map(op::Transform, t::OpTable)
+    mt = _materialize_for(t, _inputs(op, _opgranule(t)))
     _transform!(op, mt)
-    return _drop_transient(mt, transient)
+    return mt
 end
 
 """
-    apply!(op::Operation, t)
+    map!(op::Transform, t)
 
-In-place version of [`apply`](@ref). Requires `t` to be a materialized, mutable
+In-place version of [`map`](@ref). Requires `t` to be a materialized, mutable
 table (e.g. `DataFrame`, `Table`, `PartitionedTable`, `NamedTuple` of vectors).
 """
-function apply!(op::Operation, t::OpTable)
-    isempty(outputs(op)) &&
-        throw(ArgumentError("$(typeof(op)) is a filter; use `filter!(op, t)`."))
-    g = _opgranule(t)
-    _validate(t, _input_names(op, g))
+function Base.map!(op::Transform, t::OpTable)
+    _validate(t, _input_names(op, _opgranule(t)))
     _transform!(op, t)
     return t
 end
 
 """
-    filter(op::Operation, t)
+    filter(op::Filter, t)
 
 Filter rows of `t` with filter `op`, returning a new table with only the rows
 where `op`'s predicate holds. For a lazy `H5Table`/`PartitionedH5Table`, missing
-input columns are auto-pulled and then dropped. See [`filter!`](@ref).
+input columns are auto-pulled and kept. See [`filter!`](@ref).
 """
-function Base.filter(op::Operation, t::OpTable)
-    isempty(outputs(op)) ||
-        throw(ArgumentError("$(typeof(op)) is a transform; use `apply(op, t)`."))
-    g = _opgranule(t)
-    mt, transient = _materialize_for(t, inputs(op, g))
-    out = _filter_rows(op, mt)
-    return _drop_transient(out, transient)
+function Base.filter(op::Filter, t::OpTable)
+    mt = _materialize_for(t, _inputs(op, _opgranule(t)))
+    return _filter_rows(op, mt)
 end
 
 """
-    filter!(op::Operation, t)
+    filter!(op::Filter, t)
 
 In-place version of [`filter`](@ref). Requires `t` to be a materialized, mutable
 table.
 """
-function Base.filter!(op::Operation, t::OpTable)
-    isempty(outputs(op)) ||
-        throw(ArgumentError("$(typeof(op)) is a transform; use `apply!(op, t)`."))
-    g = _opgranule(t)
-    _validate(t, _input_names(op, g))
+function Base.filter!(op::Filter, t::OpTable)
+    _validate(t, _input_names(op, _opgranule(t)))
     for c in _containers(t)
-        m = mask(op, c)
+        m = _mask(op, c)
         drop = findall(!, m)
         for name in _colnames(c)
             deleteat!(Tables.getcolumn(c, name), drop)
@@ -259,25 +236,25 @@ function Base.filter!(op::Operation, t::OpTable)
 end
 
 # Non-mutating row filter that preserves container structure.
-function _filter_rows(op::Operation, t::Table)
+function _filter_rows(op::Filter, t::Table)
     nt = _table(t)
-    m = mask(op, nt)
+    m = _mask(op, nt)
     Table(map(c -> c[m], nt), _granule(t))
 end
-function _filter_rows(op::Operation, t::PartitionedTable)
+function _filter_rows(op::Filter, t::PartitionedTable)
     parts = map(t.tables) do nt
-        m = mask(op, nt)
+        m = _mask(op, nt)
         map(c -> c[m], nt)
     end
     PartitionedTable(parts, _granule(t))
 end
-function _filter_rows(op::Operation, t::DataFrame)
-    m = mask(op, t)
+function _filter_rows(op::Filter, t::DataFrame)
+    m = _mask(op, t)
     t[m, :]
 end
 
 # Apply a transform to every mutable container of `t`.
-_transform!(op::Operation, t) = (foreach(c -> _run!(op, c), _containers(t)); t)
+_transform!(op::Transform, t) = (foreach(c -> _run!(op, c), _containers(t)); t)
 
 # ─── generic operations ───────────────────────────────────────────────────────
 # Product-bound operations (TopexToWGS84, SaturationCorrect, ICESatQuality) live
@@ -291,12 +268,12 @@ Filter: keep rows whose `:longitude`/`:latitude` fall within `extent` (an
 `Extents.Extent` with `X` and `Y` bounds). Generic — applies to any granule.
 Generalizes [`in_bbox`](@ref).
 """
-struct InExtent <: Operation
+struct InExtent <: Filter
     extent::Extent
 end
 InExtent(; kwargs...) = InExtent(Extent(; kwargs...))
-inputs(::InExtent, granule) = _point_variables(granule)[1:2]
-function mask(op::InExtent, cols)
+_inputs(::InExtent, granule) = _point_variables(granule)[1:2]
+function _mask(op::InExtent, cols)
     x = Tables.getcolumn(cols, :longitude)
     y = Tables.getcolumn(cols, :latitude)
     xmin, xmax = op.extent.X
